refactor(agents): extract follow-up baseline materialization from cook.rs into cook_baseline
Third and largest carve of the cook.rs god-file (after cook_budget and
cook_promotion). Lifts the process-local baseline-materialization machinery
into a new cook_baseline module (~440 lines):
- CookFollowUpBaseline (owns a detached git worktree, cleaned up in its Drop)
- DerivedCookBaselineCapability (the non-serializable controller-validated
  authority derived from the baseline)
- materialize_initial_candidate_baseline / materialize_follow_up_baseline
- cook_attempt_harvest_context + the private git_output/git_output_with_env
  helpers (this cluster is their only user)

Pure move, no behavior change. Visibility: the fns/types cook.rs (and
cook_recipe/execution) call cross-module are pub(crate); the Drop impl and git
helpers move with the struct. Repointed the two external references
(execution.rs, cook_recipe.rs) from super::cook:: to super::cook_baseline::.
cook.rs imports back what it still calls; cook.rs tests import git_output and
access the now-pub(crate) baseline fields. Trimmed cook.rs's now-unused
Command/Digest/normalize_promotion_patch imports (moved Command/Digest into the
test module where they're still used).

cook.rs 3142 -> 2705; combined with the budget + promotion carves, cook.rs is
now 2705 vs the original 3575 (-870).

Verified: homeboy-agents --lib + --tests clean, no unused-import warnings. The
three baseline tests (follow_up_baseline_is_clean..., follow_up_baseline_refuses...,
orphaned_recipe_materializes_once...) pass in isolation.
