+++ crates/homeboy-agents/src/agent_task_service/mod.rs
@@ -6,6 +6,7 @@
 //! `crate::agent_task_service::*` unchanged.
 
 mod cook;
+mod cook_baseline;
 mod cook_budget;
 mod cook_promotion;
 mod cook_recipe;
@@ -15,6 +16,7 @@ mod reconcile;
 mod status_support;
 
 pub use cook::*;
+pub use cook_baseline::*;
 pub use cook_budget::*;
 pub use cook_promotion::*;
 pub use cook_recipe::*;
